Reorganized fonts and css folders, removing outdated deps
Removed pure-min.css, and the extraneous Oswald font.  Also moved the open-sans.css glue file into the fonts folder as that makes a little more sense, and got rid of the css folder entirely.  The goal now is that component-specific css will live next to the relevant component, and global css will live in App.css.
I didn't touch any of the semantic dependencies, and that's still doing the styles for the dashboard cards for now.