Created new loginmodal component for login routing
Created a new component to be displayed in a modal on top of any existing page when the ?login query is present.
Recreated (a basic) version of the default login modal provided by uport-connect
Hacked around the uriHandler and closeUriHandler properties of Connect to capture the short period between a login event happening and the credentials being parsed and verified, so that we can display a spinner.  This should probably be possible within uport-connect, so I may try to fix that.
After closeUriHandler fires, the spinner is displayed in the login modal.  Maybe will switch this to the main page, stay tuned.