Moved modal and spinner management to redux
Changed the LoginButton to dispatch a showSpinner() action instead of navigating to /?login.  This both solves the bug where redirect info was lost with the previous LoginModal, and allows more complicated navigation involving spinners and modals.  The modal state is read by the ModalWrapper component in components/wrappers, and any navigation event clears the modal.